modify; apply infer on all images in valid_data

+++ src/remove_background/infer.py
@@ -5,8 +5,9 @@ from torch import nn
 import cv2
 import numpy as np
 
-from .model.model_unet_pp import build_unetplusplus
+from .model.model_unet_pp import build_model
 from ..utils.arg_parser import get_parser
+from ..utils.misc import resize_with_aspect_ratio
 
 # to download model's weights, execute the following command:
 # scp <username>@<ip>:/home/ubuntu/projects/MagazineCrop/src/remove_background/checkpoints/<model_name> ./src/remove_background/checkpoints/
@@ -16,41 +17,62 @@ if __name__ == "__main__":
     parser = get_parser()
     args = parser.parse_args()
 
-    path_to_images = os.path.join(os.getcwd(), "data", "example")
-
-    image = cv2.imread(
-        os.path.join(path_to_images, args.image_name), cv2.IMREAD_GRAYSCALE
+    new_width, new_height = 1024, 1024
+    dir_names = [
+        "B6855",
+        "B6960",
+        "B6995",
+        "C2789",
+        "C2797",
+        "C2811",
+        "C2926",
+        "C3909",
+        "C3920",
+    ]
+    path_to_model = os.path.join(
+        os.getcwd(), "src", "remove_background", "checkpoints", args.model_name
     )
-    # image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-    resized_image = cv2.resize(image, (2048, 2048))
-    eh_image = cv2.equalizeHist(resized_image)
 
-    # cv2.imshow("original image", cv2.cvtColor(resized_image, cv2.COLOR_RGB2BGR))
-    cv2.imshow("original image", resized_image)
-    cv2.imshow("thresholded image", eh_image)
-
-    model = build_unetplusplus(number_of_classes=1)
-    model.load_state_dict(
-        torch.load(
-            os.path.join(args.path_to_model_dir, args.model_name), weights_only=True
-        )
-    )
+    model = build_model(number_of_classes=1)
+    model.load_state_dict(torch.load(path_to_model, weights_only=True))
     model.to(args.device)
     model.eval()
 
-    with torch.no_grad():
-        in_image = torch.tensor(eh_image).unsqueeze(0).unsqueeze(-1).float() / 255.0
-        in_image = in_image.to(args.device)
-        logits = model(in_image)
-        is_fg_prob = logits.sigmoid().squeeze().cpu().numpy()
-        fg_mask = is_fg_prob >= 0.4
+    for dir_name in dir_names:
+        path_to_dir = os.path.join(
+            os.getcwd(), "data", "valid_data", "scanned", "images", dir_name
+        )
+        if not os.path.isdir(path_to_dir):
+            continue
 
-        print(np.unique(fg_mask))
+        for image_name in os.listdir(path_to_dir):
+            if not image_name.endswith(".tif"):
+                continue
 
-        masked_image = resized_image.copy()
-        masked_image[~fg_mask] = 0
+            image = cv2.imread(
+                os.path.join(path_to_dir, image_name), cv2.IMREAD_GRAYSCALE
+            )
+            resized_image = resize_with_aspect_ratio(
+                image, target_size=(new_width, new_height)
+            )
 
-        # cv2.imshow("masked image", cv2.cvtColor(masked_image, cv2.COLOR_RGB2BGR))
-        cv2.imshow("masked image", masked_image)
-        cv2.waitKey(0)
-        cv2.destroyAllWindows()
+            eh_image = cv2.equalizeHist(resized_image)
+
+            cv2.namedWindow("image", cv2.WINDOW_NORMAL)
+            cv2.resizeWindow("image", 2048, 1024)
+
+            with torch.no_grad():
+                in_image = torch.tensor(eh_image)[None, :, :, None].float() / 255.0
+                in_image = in_image.to(args.device)
+                logits = model(in_image)
+                is_fg_prob = logits.sigmoid().squeeze().cpu().numpy()
+            fg_mask = is_fg_prob >= 0.4
+
+            masked_image = resized_image.copy()
+            masked_image[~fg_mask] = 0
+
+            mask = fg_mask.astype(np.uint8) * 255
+
+            cv2.imshow("image", cv2.hconcat([eh_image, mask, masked_image]))
+            cv2.waitKey(0)
+            cv2.destroyAllWindows()
